webui/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "roblox-agency-ui"
version = "0.1.0"
description = "Web UI control plane for the Roblox Agency autonomous build system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "gitpython>=3.1.43",
    "apscheduler>=3.10.4",
    "watchdog>=4.0.1",
    "anthropic>=0.28.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "pydantic>=2.7.0",
    "markdown-it-py>=3.0.0",
    "croniter>=2.0.5",
]

[project.scripts]
agency = "webui.bin.agency:main"

[tool.setuptools.packages.find]
where = [".."]
include = ["webui*"]

[tool.setuptools.package-data]
"webui" = ["client/dist/**/*"]
